feat(grid): Add support for field as string array
If we have a nested objects, then the field
should be defined as path (array of string).
This way we will make sure that the raw value
will be properly calculated.
Also, with this improvement we can remove some value getters from the configuration where path was used
in order to get the value.