feat(grid): Use element width for position calculations
On the components ngrx-grid-header-item and
ngrx-dynamic-grid-header-item we do not
apply the defined width.
This width is applied on upper level, on parent
components. In order to make sure that
the right column width will be used, and that the test
will not depended on the screen where it is executed
we will use elementWidth instead of column.width.