ci(update): update build-deploy github action
Configure the workflow_dispatch event, to enable triggering of the
workflow manually.
Define a deploy as a separate workflow, that will be triggered on a
successful build and when building master branch.
Adjust the configuration of the deploy step (remove branch option,
because the default value is gh-pages, the same is true for token
option).